made seperate file for cal and updated gitignore
