+++ src/main/bluesky/parser.py
@@ -1,14 +1,14 @@
 import pandas as pd
-import math
 import sys
 from bluesky.__main__ import main
-import numpy as np
-from concurrent.futures import ThreadPoolExecutor
+from calculator import calculate_bearing
+from calculator import calculate_distance
+from calculator import calculate_new_point
 
 flights = pd.read_csv("Flights.csv")
 
-# Load airport data from your big airport file into a DataFrame
-airport_data = pd.read_csv("airports.csv")  # Replace with the actual filename
+# Load airport data from your airport file into a DataFrame
+airport_data = pd.read_csv("airports.csv")
 
 
 def get_airport_info(code, info_type):
@@ -22,7 +22,6 @@ def get_airport_info(code, info_type):
             return lat
         elif info_type == "LON":
             return lon
-
     return None  # Airport code not found in the airport data
 
 
@@ -61,70 +60,6 @@ def airlines():
     return providers
 
 
-def calculate_bearing(lat1, lon1, lat2, lon2):
-    # Convert latitude and longitude from degrees to radians
-    lat1 = math.radians(lat1)
-    lon1 = math.radians(lon1)
-    lat2 = math.radians(lat2)
-    lon2 = math.radians(lon2)
-
-    # Calculate bearing
-    y = math.sin(lon2 - lon1) * math.cos(lat2)
-    x = math.cos(lat1) * math.sin(lat2) - math.sin(lat1) * math.cos(lat2) * math.cos(lon2 - lon1)
-    theta = math.atan2(y, x)
-    bearing = (theta * 180 / math.pi + 360) % 360  # in degrees
-    return round(0 if np.isnan(bearing) else bearing, 2)
-
-
-def calculate_distance(lat1, lon1, lat2, lon2):
-    # Radius of the Earth in meters
-    R = 6371e3
-
-    # Convert latitude and longitude from degrees to radians
-    latOrig = math.radians(lat1)
-    latDest = math.radians(lat2)
-    destLatMinOrigLat = math.radians(lat2 - lat1)
-    destLonMinOrigLon = math.radians(lon2 - lon1)
-
-    # Haversine formula
-    a = math.sin(destLatMinOrigLat/2) * math.sin(destLatMinOrigLat/2) + \
-        math.cos(latOrig) * math.cos(latDest) * \
-        math.sin(destLonMinOrigLon/2) * math.sin(destLonMinOrigLon/2)
-    c = 2 * math.atan2(math.sqrt(a), math.sqrt(1-a))
-
-    # Calculate the distance in meters
-    d = R * c
-
-    return 50000 if np.isnan(d) else d
-
-
-def calculate_new_point(lat1, lon1, lat2, lon2, f, delta):
-    # Convert latitude and longitude from degrees to radians
-    lat1 = math.radians(lat1)
-    lon1 = math.radians(lon1)
-    lat2 = math.radians(lat2)
-    lon2 = math.radians(lon2)
-
-    # Calculate a and b
-    a = math.sin((1 - f) * delta) / math.sin(delta)
-    b = math.sin(f * delta) / math.sin(delta)
-
-    # Calculate x, y, and z
-    x = a * math.cos(lat1) * math.cos(lon1) + b * math.cos(lat2) * math.cos(lon2)
-    y = a * math.cos(lat1) * math.sin(lon1) + b * math.cos(lat2) * math.sin(lon2)
-    z = a * math.sin(lat1) + b * math.sin(lat2)
-
-    # Calculate the new latitude (φi) and longitude (λi)
-    phi_i = math.atan2(z, math.sqrt(x**2 + y**2))
-    lambda_i = math.atan2(y, x)
-
-    # Convert back to degrees
-    phi_i = math.degrees(phi_i)
-    lambda_i = math.degrees(lambda_i)
-
-    return phi_i, lambda_i
-
-
 def write_scene_file(filename):
     with open(filename, 'w') as file:
         for x in range(flights.shape[0]):
@@ -142,7 +77,7 @@ def write_scene_file(filename):
                     f"00:00:00.00>CRE {flightCallSign.iloc[x]} {flightType.iloc[x]} {new_lat} {new_lon} {brng} "
                     f"FL{int(flightAltitude.iloc[x])} {flightSpeed.iloc[x]}\n"
                     f"00:00:00.00>DEST {flightCallSign.iloc[x]} {flightDest.iloc[x]}\n"
-                    f"00:00:00.00>{flightCallSign.iloc[x]} AT {flightDest.iloc[x]} DO DEL {flightCallSign.iloc[x]}\n"
+                    f"00:00:00.00>{flightCallSign.iloc[x]} ATDIST {flightDest.iloc[x]} DO DEL {flightCallSign.iloc[x]}\n"
                 )
                 file.write(scenetext)
