+++ src/main/bluesky/scenefile.scn
@@ -340,9 +340,9 @@
 00:00:00.00>CRE EZS1041 A319 51.83411117251387 5.06971676494071 338.51 FL220 402
 00:00:00.00>DEST EZS1041 EHAM
 00:00:00.00>EZS1041 AT EHAM DO DEL EZS1041
-00:00:00.00>CRE KLM446 A332 nan nan 0 FL260 475
-00:00:00.00>DEST KLM446 EHAM
-00:00:00.00>KLM446 AT EHAM DO DEL KLM446
+00:00:00.00>CRE KLM446 A332 OKBK 0 FL260 475
+00:00:00.00>DEST KLM446 52.30860138 4.76388979
+00:00:00.00>KLM446 ATDIST 52.30860138 4.76388979 500 DEL KLM446
 00:00:00.00>CRE KLM1178 E190 52.69796725238009 5.326328013025931 221.55 FL380 448
 00:00:00.00>DEST KLM1178 EHAM
 00:00:00.00>KLM1178 AT EHAM DO DEL KLM1178
@@ -2200,9 +2200,9 @@
 00:00:00.00>CRE KLM20H B738 51.85266539439037 4.354611461006332 28.72 FL340 448
 00:00:00.00>DEST KLM20H EHAM
 00:00:00.00>KLM20H AT EHAM DO DEL KLM20H
-00:00:00.00>CRE BLADE03 F16 nan nan 0 FL220 486
-00:00:00.00>DEST BLADE03 EHLW
-00:00:00.00>BLADE03 AT EHLW DO DEL BLADE03
+00:00:00.00>CRE BLADE03 F16 EHVL 0 FL220 486
+00:00:00.00>DEST BLADE03 53.22859955 5.760560036
+00:00:00.00>BLADE03 ATDIST 53.22859955 5.760560036 500 DEL BLADE03
 00:00:00.00>CRE VIPER01 F16 nan nan 0 FL60 517
 00:00:00.00>DEST VIPER01 XTMC
 00:00:00.00>VIPER01 AT XTMC DO DEL VIPER01
@@ -3808,9 +3808,9 @@
 00:00:00.00>CRE TRA5108 B738 50.96395080730971 5.076954974628603 20.87 FL360 460
 00:00:00.00>DEST TRA5108 EHEH
 00:00:00.00>TRA5108 AT EHEH DO DEL TRA5108
-00:00:00.00>CRE GAFE22A TOR nan nan 0 FL190 420
-00:00:00.00>DEST GAFE22A ETSB
-00:00:00.00>GAFE22A AT ETSB DO DEL GAFE22A
+00:00:00.00>CRE GAFE22A TOR EHVL 0 FL190 420
+00:00:00.00>DEST GAFE22A 50.17380142 7.063330174
+00:00:00.00>GAFE22A ATDIST 50.17380142 7.063330174 500 DEL GAFE22A
 00:00:00.00>CRE KLM99U E190 54.47557902363141 -5.018622638411647 286.4 FL380 450
 00:00:00.00>DEST KLM99U EGAC
 00:00:00.00>KLM99U AT EGAC DO DEL KLM99U
@@ -3829,9 +3829,9 @@
 00:00:00.00>CRE SAS1549 A321 52.62680957564006 5.423717311856476 231.9 FL260 450
 00:00:00.00>DEST SAS1549 EHAM
 00:00:00.00>SAS1549 AT EHAM DO DEL SAS1549
-00:00:00.00>CRE GAFE22B TOR nan nan 0 FL190 420
-00:00:00.00>DEST GAFE22B ETSB
-00:00:00.00>GAFE22B AT ETSB DO DEL GAFE22B
+00:00:00.00>CRE GAFE22B TOR EHVL 0 FL190 420
+00:00:00.00>DEST GAFE22B 50.17380142 7.063330174
+00:00:00.00>GAFE22B ATDIST 50.17380142 7.063330174 500 DEL GAFE22B
 00:00:00.00>CRE TRA6223 B738 37.16404588082926 -4.278753007571208 199.87 FL370 446
 00:00:00.00>DEST TRA6223 LEMG
 00:00:00.00>TRA6223 AT LEMG DO DEL TRA6223
